Fully ported
This commit basically signals my original OpenGL 3.3 test (not published) being fully ported to Vulkan.

3 lights, full per-pixel, single pass, with "soft" cubemap distance shadows.

The shadows are being rendered in a separate thread, when done along with the rest of the rendering, performance was about 40FPS slower than OpenGL... In a separate thread, it's about 60FPS fast.
Feels like there is an issue in there, I'll have to investigate that.